+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2026-03-09
+
+### Added
+
+- **Frontend**: New `AddWordModal.tsx` React Component to create a UI for adding words into the table.
+- **Frontend**: New `wordSchema.ts` Zod object to perform input validation on `AddWordModal.tsx` React Component.
+
+### Changed
+
+- **Frontend**: Enhanced `page.tsx` with `AddWordModal` React Component.
+
 ## [0.2.1] - 2026-03-05
 
 ### Added